requirements.txt
asgiref==3.3.1
Django==3.1.7
pytz==2021.1
sqlparse==0.4.1
psycopg2>=2.8
redis>=3.5
celery>=5.0
django-mptt==0.13.4
pillow
djangorestframework
djangorestframework-simplejwt
django-rest-swagger==2.2.0
drf-yasg